requirements.txt
altair==5.1.2
annotated-types==0.6.0
attrs==23.1.0
blinker==1.6.3
blis==0.7.11
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.3
cymem==2.0.8
gitdb==4.0.11
GitPython==3.1.40
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kaleido==0.2.1
langcodes==3.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
murmurhash==1.0.10
numpy==1.26.1
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
plotly==5.18.0
preshed==3.0.9
protobuf==4.24.4
pt-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_lg-3.7.0/pt_core_news_lg-3.7.0-py3-none-any.whl#sha256=4eed7801647c39ced078eb325d0ff8eb2ea5d4945501b8ee898dced85c686717
pt-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.7.0/pt_core_news_sm-3.7.0-py3-none-any.whl#sha256=a3a8eed42c600af17c70ed3c8720ffe2879b5353055e66fe0c998306329052fd
pyarrow==13.0.0
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.10.6
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.4.8
streamlit==1.28.0
streamlit-ace==0.1.1
streamlit-code-editor==0.1.10
streamlit-elements==0.1.0
tabulate==0.9.0
tenacity==8.2.3
thinc==8.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
validators==0.22.0
vanna==0.0.27
wasabi==1.1.2
watchdog==3.0.0
weasel==0.3.3
zipp==3.17.0
